#17110 Remove dependency on service test utils from platform tests
