+++ src/vs/platform/storage/test/storageService.test.ts
@@ -7,15 +7,24 @@
 
 import * as assert from 'assert';
 import { clone } from 'vs/base/common/objects';
+import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { StorageScope } from 'vs/platform/storage/common/storage';
-import { TestContextService, TestWorkspace } from 'vs/test/utils/servicesTestUtils';
+import { IWorkspaceContextService, WorkspaceContextService } from 'vs/platform/workspace/common/workspace';
 import { StorageService, InMemoryLocalStorage } from 'vs/platform/storage/common/storageService';
+import { TestWorkspace } from 'vs/platform/workspace/test/common/testWorkspace';
 
 suite('Workbench StorageSevice', () => {
 
+	let contextService, instantiationService;
+
+	setup(() => {
+		instantiationService = new TestInstantiationService();
+		contextService = instantiationService.stub(IWorkspaceContextService, WorkspaceContextService);
+		instantiationService.stub(IWorkspaceContextService, 'getWorkspace', TestWorkspace);
+	});
+
 	test('Swap Data with undefined default value', () => {
-		let context = new TestContextService();
-		let s = new StorageService(new InMemoryLocalStorage(), null, context);
+		let s = new StorageService(new InMemoryLocalStorage(), null, contextService);
 
 		s.swap('Monaco.IDE.Core.Storage.Test.swap', 'foobar', 'barfoo');
 		assert.strictEqual('foobar', s.get('Monaco.IDE.Core.Storage.Test.swap'));
@@ -26,8 +35,7 @@ suite('Workbench StorageSevice', () => {
 	});
 
 	test('Remove Data', () => {
-		let context = new TestContextService();
-		let s = new StorageService(new InMemoryLocalStorage(), null, context);
+		let s = new StorageService(new InMemoryLocalStorage(), null, contextService);
 		s.store('Monaco.IDE.Core.Storage.Test.remove', 'foobar');
 		assert.strictEqual('foobar', s.get('Monaco.IDE.Core.Storage.Test.remove'));
 
@@ -36,8 +44,7 @@ suite('Workbench StorageSevice', () => {
 	});
 
 	test('Get Data, Integer, Boolean', () => {
-		let context = new TestContextService();
-		let s = new StorageService(new InMemoryLocalStorage(), null, context);
+		let s = new StorageService(new InMemoryLocalStorage(), null, contextService);
 
 		assert.strictEqual(s.get('Monaco.IDE.Core.Storage.Test.get', StorageScope.GLOBAL, 'foobar'), 'foobar');
 		assert.strictEqual(s.get('Monaco.IDE.Core.Storage.Test.get', StorageScope.GLOBAL, ''), '');
@@ -71,15 +78,14 @@ suite('Workbench StorageSevice', () => {
 
 	test('StorageSevice cleans up when workspace changes', () => {
 		let storageImpl = new InMemoryLocalStorage();
-		let context = new TestContextService();
-		let s = new StorageService(storageImpl, null, context);
+		let s = new StorageService(storageImpl, null, contextService);
 
 		s.store('key1', 'foobar');
 		s.store('key2', 'something');
 		s.store('wkey1', 'foo', StorageScope.WORKSPACE);
 		s.store('wkey2', 'foo2', StorageScope.WORKSPACE);
 
-		s = new StorageService(storageImpl, null, context);
+		s = new StorageService(storageImpl, null, contextService);
 
 		assert.strictEqual(s.get('key1', StorageScope.GLOBAL), 'foobar');
 		assert.strictEqual(s.get('key1', StorageScope.WORKSPACE, null), null);
@@ -90,8 +96,8 @@ suite('Workbench StorageSevice', () => {
 
 		let ws: any = clone(TestWorkspace);
 		ws.uid = new Date().getTime() + 100;
-		context = new TestContextService(ws);
-		s = new StorageService(storageImpl, null, context);
+		instantiationService.stub(IWorkspaceContextService, 'getWorkspace', ws);
+		s = new StorageService(storageImpl, null, contextService);
 
 		assert.strictEqual(s.get('key1', StorageScope.GLOBAL), 'foobar');
 		assert.strictEqual(s.get('key1', StorageScope.WORKSPACE, null), null);
