+++ src/vs/test/utils/servicesTestUtils.ts
@@ -11,7 +11,6 @@ import { Promise, TPromise } from 'vs/base/common/winjs.base';
 import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { EventEmitter } from 'vs/base/common/eventEmitter';
 import * as paths from 'vs/base/common/paths';
-import * as assert from 'assert';
 import URI from 'vs/base/common/uri';
 import { ITelemetryService, NullTelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { StorageService, InMemoryLocalStorage } from 'vs/platform/storage/common/storageService';
@@ -51,12 +50,9 @@ import { HistoryService } from 'vs/workbench/services/history/browser/history';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { TestConfigurationService } from 'vs/platform/configuration/test/common/testConfigurationService';
 import { IWindowsService, IWindowService } from 'vs/platform/windows/common/windows';
+import { TestWorkspace as Workspace } from 'vs/platform/workspace/test/common/testWorkspace';
 
-export const TestWorkspace: IWorkspace = {
-	resource: URI.file('C:\\testWorkspace'),
-	name: 'Test Workspace',
-	uid: Date.now()
-};
+export const TestWorkspace: IWorkspace = Workspace;
 
 export function createFileInput(instantiationService: IInstantiationService, resource: URI): FileEditorInput {
 	return instantiationService.createInstance(FileEditorInput, resource, void 0);
@@ -793,16 +789,6 @@ export class TestLifecycleService implements ILifecycleService {
 export function createMockModelService(instantiationService: TestInstantiationService): IModelService {
 	instantiationService.stub(IConfigurationService, new TestConfigurationService());
 	return instantiationService.createInstance(ModelServiceImpl);
-}
-
-export function onError(error: Error, done: () => void): void {
-	assert.fail(error);
-
-	done();
-}
-
-export function toResource(path) {
-	return URI.file(paths.join('C:\\', new Buffer(this.test.fullTitle()).toString('base64'), path));
 }
 
 export class TestWindowsService implements IWindowsService {
