+++ src/vs/workbench/services/files/test/node/fileService.test.ts
@@ -19,7 +19,7 @@ import uuid = require('vs/base/common/uuid');
 import extfs = require('vs/base/node/extfs');
 import encodingLib = require('vs/base/node/encoding');
 import utils = require('vs/workbench/services/files/test/node/utils');
-import { onError } from 'vs/test/utils/servicesTestUtils';
+import { onError } from 'vs/base/test/common/utils';
 
 suite('FileService', () => {
 	let events: utils.TestEventService;
